project/requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
gunicorn==21.0.1
pydantic-settings==2.1.0
asyncpg==0.29.0
tortoise-orm==0.21.5
aerich==0.7.2
httpx==0.26.0
pytest==7.4.4
python-multipart==0.0.9
pyjwt==2.9.0
argon2-cffi==23.1.0